dashboard-api/requirements.txt
fastapi==0.115.1
uvicorn==0.30.0
psycopg2-binary==2.9.9